+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.25.41
-* **Try again** added in connecting status
+* `Try again` added in connecting status
 * Translations updated
 * Waiting for network indicator improved